requirements/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysql-connector-python==8.3.0
requests==2.28.1
